+++ docs/src/markdown/about/changelog.md
@@ -3,7 +3,7 @@
 ## 4.6.0
 
 -   **NEW**: Add new `rounding` option to control rounding modes in `to_string()`, `to_dict()`, `coords()`, `alpha()`,
-    and `get()`.
+    and `get()`. New rounding modes are: `sigfig` and `decimal` which are in addition to the default `digits` mode.
 -   **FIX**: Fix typing for inherited colors.
 
 ## 4.5.1